pyproject.toml
[tool.poetry]
name = "queue-monitor"
version = "0.1.0"
description = ""
authors = ["Mahra Rahimi <mahra.rah@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
azure-functions = "^1.8.0"
azure-monitor-opentelemetry = "^1.6.1"
azure-storage-queue = "^12.8.1"
azure-identity = "^1.6.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.0.0"
black = "^24.8.0"
isort = "^5.13.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
